Set DEPNAME to libexpat for the expat library
In a previous commit (16f878bf2d6e06b3b7e4970892c132b19b8ced81)
the library .def file was updated to specify 'libexpat' as the
LIBRARY parameter. However, the cmake was not updated to reflect
that change, thus MSVC-generated projects still bore the old
'expat' name. This made the .dlls generated to be improperly named
for what the linked binaries expected (libexpat).