Fix libpng exported module based on build type
We need to take into account the current build type
for the proper .def file. However in MSVC's case, since
it's a multi configuration we can't rely on the
CMAKE_BUILD_TYPE at configure time, so we tweak the
linker flags to take the proper .def file based on
the configuration.